fix segmentation fault when training on wiki-news-300d-1M.vec
After reading the last line of wiki-news-300d-1M.vec, the EOF flag was
not set so the load_embedding() function kept reading words and values.
But the array used to store the values of the vectors was full (because
all the lines have been read, and the array has a fixed size) so it was
overwriting to memory *outside* the array, with the value 0 (because
EOF is reached, so read_float() returns 0).
The overwriting sets the first saved words to 0 (or NULL because they
are pointers) so when the function free() was called to free the memory,
it was trying to dereference a NULL pointer, hence the segmentation
fault.

I modified the stop condition to make sure the program can't read values
and overwrite memory when EOF is reached.