Remove duplicate code in topk_binary.c
The file topk_binary.c uses a custom version of the function to load
binary vectors from a file (like the one used in similarity_binary.c).
For this reason, a new function was created and the util functions
required to make it work were copied from hashtab.c and file_process.c
(this was the easiest solution found during the devlopment). Remove
those duplicate copied functions in topk_binary.c