Rewrite load_binary_vectors to load all the vectors (for topk)
To use the function load_binary_vectors() into topk_binary.c (and not
redefine a custome version in this file), load all the vectors from the
embedding file instead of only loading the ones already in the hashtab
(this was the default behavior because of similarity_binary.c which
loads only the vectors of words in the evaluation datasets).  Moreover,
also replace the global variables in topk_binary.c with local variables.