Refactor topk_binary.c to move the variable `words` into hashtab.c
The array used to know a word given its index was declared inside
topk_binary.c (because it is only used for this program). Move it to
hashtab.c, because it is more logical to put it there. Have to modify
some functions like `add_word()` and `load_binary_vectors()`.