Add flag in load_binary_vectors to not load all vectors from file
The programs `topk_binary` and `similarity_binary` are slightly
different. `topk_binary` needs to have all the vectors from the
embedding file to compute the closest neighbors, and it also needs the
index->word array from hashtab.c. `similarity_binary` only needs the
vectors for the words in the evaluation datasets. Instead of having two
different functions, use a flag in `load_binary_vectors()` to choose to
load all the vectors or not, because the rest of the code (like
allocating memory for vectors or reading values from file) is common.